py-scripts/requirements.txt
tushare
pandas
snownlp
supabase
python-dotenv